1. Поправил код, приложение запускается 2. Вынес реализацию метода clean() из WebViewViewController 3. UI тесты проходят 4. Убрал конструкции assertionFailure из. кода 5. Вынес в отдельные файлы каждую структуру в ProfileService
