+++ ImageFeed/Auth/AuthViewController.swift
@@ -22,9 +22,7 @@ final class AuthViewController: UIViewController {
         if segue.identifier == ShowWebViewSegueIdentifier {
             guard
                 let webViewViewController = segue.destination as? WebViewViewController
-            else { assertionFailure("Failed to prepare for \(ShowWebViewSegueIdentifier)")
-                return
-            }
+            else { fatalError("Failed to prepare for \(ShowWebViewSegueIdentifier)") }
             let authHelper = AuthHelper()
             webViewViewController.delegate = self
             let webViewPresenter = WebViewPresenter(authHelper: authHelper)
@@ -40,7 +38,10 @@ final class AuthViewController: UIViewController {
 
 extension AuthViewController: WebViewViewControllerDelegate {
     func webViewViewController(_ vc: WebViewViewController, didAuthenticateWithCode code: String) {
-        delegate?.authViewController(self, didAuthenticateWithCode: code)
+        dismiss(animated: true) { [weak self] in
+            guard let self = self else { return }
+            delegate?.authViewController(self, didAuthenticateWithCode: code)
+        }
     }
     
     func webViewViewControllerDidCancel(_ vc: WebViewViewController) {
