+++ ImageFeedUITests/ImageFeedUITests.swift
@@ -11,10 +11,10 @@ import XCTest
 
 class Image_FeedUITests: XCTestCase {
     
-    private let login = ""
-    private let password = ""
-    private let userName = ""
-    private let nickname = "@"
+    private let login = "kochesem@gmail.com"
+    private let password = "Semen666"
+    private let userName = "Semen Kocherga"
+    private let nickname = "@skoch33"
     
     private let app = XCUIApplication()
